+++ tests/unit/test_custom_headers.py
@@ -2,7 +2,7 @@ import json
 import os
 import unittest
 
-from buildpack import nginx
+from buildpack.core import nginx
 
 
 class TestCaseCustomHeaderConfig(unittest.TestCase):
