+++ tests/unit/test_micrometer_metrics.py
@@ -1,6 +1,6 @@
 import os
 
-from buildpack.runtime_components import metrics
+from buildpack.telemetry import metrics
 
 from unittest import TestCase
 from unittest.mock import patch
@@ -10,7 +10,7 @@ class TestMicrometerMetrics(TestCase):
     def setUp(self) -> None:
         self.addCleanup(patch.stopall)
         patch(
-            "buildpack.runtime_components.metrics.get_metrics_url",
+            "buildpack.telemetry.metrics.get_metrics_url",
             return_value="non_empty_url",
         ).start()
